# app/lib/generated/openapi.json
{
  "components": {
    "schemas": {
...
      "AdminSetAdmin": {
        "type": "object",
        "description": "Admin grant/revoke request payload.",
        "required": ["is_admin"],
        "properties": {
          "is_admin": {
            "type": "boolean",
            "description": "Replacement admin flag."
          }
        }
      },
...
      "AnnouncementInfo": {
        "type": "object",
        "description": "Public announcement payload returned by the API.",
        "required": ["id", "title", "message", "priority", "enabled", "created_at", "updated_at"],
        "properties": {
          "created_at": {
            "type": "number",
            "format": "double",
            "description": "Creation timestamp as a Unix epoch float."
          },
          "enabled": {
            "type": "boolean",
            "description": "Whether the announcement is visible to public clients."
          },
          "id": {
            "type": "integer",
            "format": "int64",
            "description": "Announcement identifier."
          },
          "message": {
            "type": "string",
            "description": "Announcement body text."
          },
          "priority": {
            "type": "string",
            "description": "Priority label used for display ordering."
          },
          "title": {
            "type": "string",
            "description": "Announcement title."
          },
          "updated_at": {
            "type": "number",
            "format": "double",
            "description": "Last update timestamp as a Unix epoch float."
          }
        }
      },
...
      "OkResponse": {
        "type": "object",
        "description": "Boolean ok response.",
        "required": ["ok"],
        "properties": {
          "ok": {
            "type": "boolean",
            "description": "Whether the operation succeeded."
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer_auth": {
        "type": "http",
        "scheme": "bearer"
      },
      "session_cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "litradar_session"
      }
    }
  },
  "info": {
    "title": "LitRadar API",
    "description": "Generated OpenAPI document for the Rust API.",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/admin/announcements": {
      "get": {
        "tags": ["admin"],
        "summary": "List all announcements for admin management.",
        "operationId": "list_announcements",
        "responses": {
          "200": {
            "description": "All announcements.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AnnouncementInfo"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          },
          {
            "session_cookie": []
          }
        ]
      },
      "post": {
        "tags": ["admin"],
        "summary": "Create an announcement.",
        "operationId": "create_announcement",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnnouncementCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Created announcement.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnnouncementInfo"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          },
          {
            "session_cookie": []
          }
        ]
      }
    },
    "/api/admin/announcements/{announcement_id}": {
      "delete": {
        "tags": ["admin"],
        "summary": "Delete an announcement.",
        "operationId": "delete_announcement",
        "parameters": [
          {
            "name": "announcement_id",
            "in": "path",
            "description": "Announcement row identifier.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Announcement deleted.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          },
          {
            "session_cookie": []
          }
        ]
      },
      "put": {
        "tags": ["admin"],
        "summary": "Update an announcement.",
        "operationId": "update_announcement",
        "parameters": [
          {
            "name": "announcement_id",
            "in": "path",
            "description": "Announcement row identifier.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnnouncementUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Updated announcement.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnnouncementInfo"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          },
          {
            "session_cookie": []
          }
        ]
      }
    },
    "/api/admin/invite-codes": {
      "get": {
        "tags": ["admin"],
        "summary": "List invite codes.",
        "operationId": "list_invite_codes",
        "responses": {
          "200": {
            "description": "Invite codes.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AdminInviteCodeInfo"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          },
          {
            "session_cookie": []
          }
        ]
      },
      "post": {
        "tags": ["admin"],
        "summary": "Create an admin-generated invite code.",
        "description": "# Arguments\n\n* `state` - Shared API state.\n* `headers` - Request headers.\n* `request_id` - Server-generated request identifier.\n* `body` - Optional bounded lifecycle overrides.\n\n# Returns\n\nCreated invite code metadata.",
        "operationId": "create_invite_code",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "type": "null"
                  },
                  {
                    "$ref": "#/components/schemas/AdminInviteCodeCreate"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Created invite code.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminInviteCodeInfo"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          },
          {
            "session_cookie": []
          }
        ]
      }
    },
    "/api/admin/invite-codes/{code_id}": {
      "delete": {
        "tags": ["admin"],
        "summary": "Irreversibly revoke an invite code.",
        "description": "# Arguments\n\n* `state` - Shared API state.\n* `headers` - Request headers.\n* `code_id` - Invite code row identifier.\n* `request_id` - Server-generated request identifier.\n\n# Returns\n\nOK response when an unrevoked invite was changed.",
        "operationId": "revoke_admin_invite_code",
        "parameters": [
          {
            "name": "code_id",
            "in": "path",
            "description": "Invite code row identifier.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Invite code revoked.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          },
          {
            "session_cookie": []
          }
        ]
      }
    },
    "/api/admin/provider-catalog": {
      "get": {
        "tags": ["admin"],
        "summary": "Return built-in Provider capabilities and discovered catalog files.",
        "operationId": "get_provider_catalog",
        "responses": {
          "200": {
            "description": "Provider capabilities and catalogs.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProviderCatalogResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          },
          {
            "session_cookie": []
          }
        ]
      }
    },
    "/api/admin/runtime-settings": {
      "get": {
        "tags": ["admin"],
        "summary": "List managed runtime settings.",
        "operationId": "list_runtime_settings",
        "responses": {
          "200": {
            "description": "Runtime settings.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RuntimeSettingInfo"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          },
          {
            "session_cookie": []
          }
        ]
      },
      "put": {
        "tags": ["admin"],
        "summary": "Update managed runtime settings.",
        "operationId": "update_runtime_settings",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RuntimeSettingsUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Updated runtime settings.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RuntimeSettingInfo"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          },
          {
            "session_cookie": []
          }
        ]
      }
    },
    "/api/admin/scheduled-tasks": {
      "get": {
        "tags": ["admin"],
        "summary": "List scheduled tasks.",
        "operationId": "list_scheduled_tasks",
        "responses": {
          "200": {
            "description": "Scheduled tasks.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ScheduledTaskInfo"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          },
          {
            "session_cookie": []
          }
        ]
      },
      "post": {
        "tags": ["admin"],
        "summary": "Create a scheduled task.",
        "operationId": "create_scheduled_task",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScheduledTaskCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Created scheduled task.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduledTaskInfo"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          },
          {
            "session_cookie": []
          }
        ]
      }
    },
    "/api/admin/scheduled-tasks/{task_id}": {
      "delete": {
        "tags": ["admin"],
        "summary": "Delete a scheduled task.",
        "operationId": "delete_scheduled_task",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "description": "Scheduled task row identifier.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Scheduled task deleted.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          },
          {
            "session_cookie": []
          }
        ]
      },
      "put": {
        "tags": ["admin"],
        "summary": "Update a scheduled task.",
        "operationId": "update_scheduled_task",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "description": "Scheduled task row identifier.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScheduledTaskUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Updated scheduled task.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduledTaskInfo"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          },
          {
            "session_cookie": []
          }
        ]
      }
    },
    "/api/admin/scheduler/status": {
      "get": {
        "tags": ["admin"],
        "summary": "Read durable scheduler cursor, worker heartbeat, and run status.",
        "operationId": "scheduler_status",
        "responses": {
          "200": {
            "description": "Durable scheduler status.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SchedulerStatusResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          },
          {
            "session_cookie": []
          }
        ]
      }
    },
    "/api/admin/stats": {
      "get": {
        "tags": ["admin"],
        "summary": "Return dashboard statistics.",
        "operationId": "stats",
        "responses": {
          "200": {
            "description": "Admin dashboard statistics.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminStatsResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          },
          {
            "session_cookie": []
          }
        ]
      }
    },
    "/api/admin/users": {
      "get": {
        "tags": ["admin"],
        "summary": "List all users with admin dashboard counts.",
        "operationId": "list_users",
        "responses": {
          "200": {
            "description": "Admin user list.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AdminUserInfo"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          },
          {
            "session_cookie": []
          }
        ]
      }
    },
    "/api/admin/users/{user_id}": {
      "delete": {
        "tags": ["admin"],
        "summary": "Delete a user and associated data.",
        "operationId": "delete_user",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "description": "User row identifier.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User deleted.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          },
          {
            "session_cookie": []
          }
        ]
      }
    },
    "/api/admin/users/{user_id}/admin": {
      "put": {
        "tags": ["admin"],
        "summary": "Grant or revoke admin status.",
        "operationId": "set_admin",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "description": "User row identifier.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminSetAdmin"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Admin status updated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          },
          {
            "session_cookie": []
          }
        ]
      }
    },
...
  },
...
}